Make the gr_modtool generated template compilable
The C++ code generated by gr_modtool does not compile out of the box:
it includes some markers (ex: "<+MARKERS+>") to highlight code portions
that needs to be customized. This commit just replace those markers with
relevant values, to make the code compile again with this new module
code structure. The blocks themselves, nor the tests are not expected to
work at this point.

Signed-off-by: Julien Olivain <juju@cotds.org>